allow user to change dumping flags in the deck
