+++ src/vpic/hdf5_header_info.h
@@ -2,143 +2,6 @@
 #define VPIC_HDF5_HEAD_INFO
 
 #define FIELD_ARRAY_NAME field_array
-struct field_dump_flag_t
-{
-  bool ex = true, ey = true, ez = true, div_e_err = true;
-  bool cbx = true, cby = true, cbz = true, div_b_err = true;
-  bool tcax = true, tcay = true, tcaz = true, rhob = true;
-  bool jfx = true, jfy = true, jfz = true, rhof = true;
-  bool ematx = true, ematy = true, ematz = true, nmat = true;
-  bool fmatx = true, fmaty = true, fmatz = true, cmat = true;
-  void disableE()
-  {
-    ex = false, ey = false, ez = false, div_e_err = false;
-  }
-
-  void disableCB()
-  {
-    cbx = false, cby = false, cbz = false, div_b_err = false;
-  }
-
-  void disableTCA()
-  {
-    tcax = false, tcay = false, tcaz = false, rhob = false;
-  }
-
-  void disableJF()
-  {
-    jfx = false, jfy = false, jfz = false, rhof = false;
-  }
-
-  void disableEMAT()
-  {
-    ematx = false, ematy = false, ematz = false, nmat = false;
-  }
-
-  void disableFMAT()
-  {
-    fmatx = false, fmaty = false, fmatz = false, cmat = false;
-  }
-
-  void resetToDefaults()
-  {
-    ex = true, ey = true, ez = true, div_e_err = true;
-    cbx = true, cby = true, cbz = true, div_b_err = true;
-    tcax = true, tcay = true, tcaz = true, rhob = true;
-    jfx = true, jfy = true, jfz = true, rhof = true;
-    ematx = true, ematy = true, ematz = true, nmat = true;
-    fmatx = true, fmaty = true, fmatz = true, cmat = true;
-  }
-
-  bool enabledE()
-  {
-    return ex && ey && ez;
-  }
-
-  bool enabledCB()
-  {
-    return cbx && cby && cbz;
-  }
-
-  bool enabledTCA()
-  {
-    return tcax && tcay && tcaz;
-  }
-
-  bool enabledJF()
-  {
-    return jfx && jfy && jfz;
-  }
-
-  bool enabledEMAT()
-  {
-    return ematx && ematy && ematz;
-  }
-
-  bool enabledFMAT()
-  {
-    return fmatx && fmaty && fmatz;
-  }
-};
-
-struct hydro_dump_flag_t
-{
-  bool jx = true, jy = true, jz = true, rho = true;
-  bool px = true, py = true, pz = true, ke = true;
-  bool txx = true, tyy = true, tzz = true;
-  bool tyz = true, tzx = true, txy = true;
-
-  void disableJ()
-  {
-    jx = false, jy = false, jz = false, rho = false;
-  }
-
-  void disableP()
-  {
-    px = false, py = false, pz = false, ke = false;
-  }
-
-  void disableTD() //Stress diagonal
-  {
-    txx = false, tyy = false, tzz = false;
-  }
-
-  void disableTOD() //Stress off-diagonal
-  {
-    tyz = false, tzx = false, txy = false;
-  }
-  void resetToDefaults()
-  {
-    jx = true, jy = true, jz = true, rho = true;
-    px = true, py = true, pz = true, ke = true;
-    txx = true, tyy = true, tzz = true;
-    tyz = true, tzx = true, txy = true;
-  }
-
-  bool enabledJ()
-  {
-    return jx && jy && jz;
-  }
-
-  bool enabledP()
-  {
-    return px && py && pz;
-  }
-
-  bool enabledTD()
-  {
-    return txx && tyy && tzz;
-  }
-
-  bool enabledTOD()
-  {
-    return tyz && tzx && txy;
-  }
-};
-
-// Declare vars to use
-hydro_dump_flag_t hydro_dump_flag;
-field_dump_flag_t field_dump_flag;
 
 // XML header stuff
 const char *header = "<?xml version=\"1.0\"?>\n<!DOCTYPE Xdmf SYSTEM \"Xdmf.dtd\" []>\n<Xdmf xmlns:xi=\"http://www.w3.org/2001/XInclude\" Version=\"2.0\">\n\t<Domain>\n";
